+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Please make sure to add your changes to the appropriate categories:
 - Added `Filter<T>` quadratic-phase map for `Chirp<T>`
 - Added `Filter<Timed<T, Tm>>` rectangular integrator
 - Added `Filter<Timed<T, Tm>>` dt-aware prediction
+- Added `Trapezoidal<T>` dt-aware integrator
 
 ### Changed